+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changelog
 ### Version 1.0.2
 * Added option to disable exporting images with the particle file
-* Fix for reading files with CRLF end of line sequences
+* Locate Images no longer copies files to the particle effect's parent folder.
+* Fixed reading files with CRLF end of line sequences
 * Fixed stage focus when the locate images dialog
 * Fixed issues with mice that send two touch-downs instead of one
 * Minor bugfixes and code cleanup